Get rid of proprietary vector types
I want to integrate this library into an engine and realize that
having bespoke math types is annoying. So I need to switch to a
more general interface and I chose pointers to floats.
I thought this would take 2-3 hours, 5 hours tops, but it took
a whole weekend and three evenings. Crazy stuff. I'm glad it's
done.