Fix collision
Everything was broken, because the wuzy_collider pointers coming out of
a query were not mapped to a wuzy::Collider pointer properly.
I don't know why that worked before, but this cast is not allowed
if Collider is not standard layout, which it is not, because it has
virtual methods. I don't actually need Collider to be an abstract
base, so I just removed it.